server-side/pyproject.toml
[tool.poetry]
name = "austism-assistant"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "austism_assistant"}]

[tool.poetry.dependencies]
python = "3.10.6"
django = "^4.1.7"
django-phone-field = "^1.8.1"
pillow = "^9.4.0"
djangorestframework = "^3.14.0"
django-cors-headers = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
drf-nested-routers = "^0.93.4"
tensorflow = "^2.12.0"
pandas = "^2.0.1"
django-countries = "^7.5.1"
django-language-field = "^0.0.3"
opencv-python = "^4.7.0.72"
mtcnn = "^0.1.1"
django-environ = "^0.10.0"
drf-spectacular = "^0.26.2"
django-extensions = "^3.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
